+++ app-config.js
@@ -1,6 +1,6 @@
 globalThis.WC26_CONFIG = {
-  version: "704.12.33",
-  cacheName: "wc26-build-704-12-33",
+  version: "704.13.0",
+  cacheName: "wc26-build-704-13-0",
   versionUrl: "./version.json",
   updateCheckIntervalMs: 15 * 60 * 1000
 };